Add verify route
